project_files/spiders/requirements.txt
Scrapy==2.3.0
kafka-python==2.0.1